requirements.txt
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
Pillow==10.4.0